Move extension bundle code into src/common/
This commit moves a bunch of code that is common to all WebExtensions
("chrome.*") based browser extensions into src/common/ and moves the
corresponding tests into tests/common/

This makes it easier to build the extension from scratch, as it means
that code in the same place can be treated the same way. For example:

- all code in src/common/ is built into a browserify bundle
- all code in src/chrome/lib/ is copied as-is into the extension
