+++ src/common/browser-action.js
