+++ src/common/detect-content-type.js
