+++ src/common/extension.js
