+++ src/common/hypothesis-chrome-extension.js
