+++ src/common/install.js
