+++ src/common/settings.js
