+++ src/common/sidebar-injector.js
