+++ src/common/uri-info.js
