+++ tests/common/detect-content-type-test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var detectContentType = require('../lib/detect-content-type');
+var detectContentType = require('../../src/common/detect-content-type');
 
 describe('detectContentType()', function () {
   var el;
