+++ tests/karma.config.js
@@ -20,12 +20,12 @@ module.exports = function(config) {
     // list of files / patterns to load in the browser
     files: [
       {
-        pattern: 'test/settings.json',
+        pattern: './settings.json',
         included: false,
       },
-      './lib/karma-phantomjs-polyfill.js',
-      './test/bootstrap.js',
-      './test/*.js',
+      './karma-phantomjs-polyfill.js',
+      './bootstrap.js',
+      './common/*.js',
     ],
 
     // list of files to exclude
@@ -35,8 +35,8 @@ module.exports = function(config) {
     // preprocess matching files before serving them to the browser
     // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
     preprocessors: {
-      './lib/*.js': ['browserify'],
-      './test/*.js': ['browserify'],
+      '../src/common/*.js': ['browserify'],
+      './**/*.js': ['browserify'],
     },
 
     browserify: {
