+++ tests/settings.json
